config/flash overhaul!!
- update license info
- rewrite of entire flash system to increase readability, speed, and usability (don't worry it's still really bad)
- because of that, change how debug messages are printed
- prefer long doubles in nav library
- remove legacy set_pid commands
- overhaul config system to work with new flash system; restructure sections as there are less limits now!!
- add support for wifly/auto mode drop bay..thing
- fix some misc api bugs
- fix some compilation errors
- fix bug (as usual)

only really a massive commit because I had to rewrite so many bits of code that used the old config lol

I swear I'll do something productive soon!!

+++ src/lib/nav.c
@@ -7,14 +7,14 @@
 
 #include "nav.h"
 
-static inline double toRadians(double degrees) {
+static inline long double toRadians(long double degrees) {
     return degrees * M_PI / 180.0;
 }
 
-double calculateBearing(double latA, double lonA, double latB, double lonB) {
-    double thetaA = toRadians(latA);
-    double thetaB = toRadians(latB);
-    double deltaL = toRadians(lonB - lonA);
+double calculateBearing(long double latA, long double lonA, long double latB, long double lonB) {
+    long double thetaA = toRadians(latA);
+    long double thetaB = toRadians(latB);
+    long double deltaL = toRadians(lonB - lonA);
 
     double X = cos(thetaB) * sin(deltaL);
     double Y = cos(thetaA) * sin(thetaB) - sin(thetaA) * cos(thetaB) * cos(deltaL);
@@ -27,11 +27,11 @@ double calculateBearing(double latA, double lonA, double latB, double lonB) {
     return bearing;
 }
 
-double calculateDistance(double latA, double lonA, double latB, double lonB) {
-    double thetaA = toRadians(latA);
-    double thetaB = toRadians(latB);
-    double deltaT = toRadians(latB - latA);
-    double deltaL = toRadians(lonB - lonA);
+double calculateDistance(long double latA, long double lonA, long double latB, long double lonB) {
+    long double thetaA = toRadians(latA);
+    long double thetaB = toRadians(latB);
+    long double deltaT = toRadians(latB - latA);
+    long double deltaL = toRadians(lonB - lonA);
 
     double a = sin(deltaT / 2) * sin(deltaT / 2) +
                cos(thetaA) * cos(thetaB) *
